# ids-element: memo style objects for performance improvement per-render