--- a/test/helpers/css-style-sheet-mock.js
+++ b/test/helpers/css-style-sheet-mock.js
@@ -1,0 +1,3 @@
+window.CSSStyleSheet = function CSSStyleSheet() { //eslint-disable-line
+  return { cssRules: [], replaceSync: () => '' };
+};