@rob2d 2021-07-23
Seems there were a few things I didn't account for in some newer components e.g. datagrid relies on :host re-styling so can't have a shared instance of CSSStylesheet currently to alter CSS vars (unfortunately), and our Jest tests restrict what we can do on base without losing coverage and also has some quirks on CSSStyleSheet behavior. 

There's a few approaches to that but may need to close this PR to rethink that for time sake for now.