@rob2d 2021-07-23
@tmcconechy yeah had mentioned in comment above there were quirks... managed to fix/work around a few of them so far -- it seems actually mostly because of the way Jest polyfills (or doesn't) the adoptedStylesheets and the way it's CSSStyleSheet works. So we worked around it in certain ways but then to do proper checks in our app code for that causes it to fail and read the DOM as if stylesheets are not supported since JSDOM is pretty brittle. 

Was looking into some solutions yesterday to get past the nitty gritty stuff like this without losing coverage in Jest and maybe we can have a `process.env.JEST_WORKER_ID` check in ids-element to emulate shadow DOM as worst case scenario for some of these super base level things?

For the data grid I started looking into a solution last night also to try to put the CSS var relied on somewhere besides in `:host`. These sort of edge cases would still probably be an issue with `<link>` as well. 

For this branch mainly was just seeing how much faster recycling the styles could be, and can still use basically the same approach as link tags. The link tags also should be memo'd since running on render so even though there may be some browser optimization it is probably more expensive for the object/string serializing than the browser to make the styles.

Going to return to this one but just to update on it here.