@rob2d 2021-07-28
Closing for now just due to other priorities on the sprint or smaller side tasks outside of this at the moment to look at since it's not a critical issue, but it should be helpful for us to memo our style objects after seeing the practical perf gain (either way whether we're using `link` or `style` attrib, it should be the same). 

Maybe in the meantime on new component PRs in retrospect for reviews it can be good for us to keep an eye out for style via JS on "host" with any vars specifically since these cases throw a small monkey wrench to untangle (not a bad practice in itself, just based on if we want to be able to share resources between similar components with the same theme context).