@rob2d 2021-08-09
just to add a thought on this one for whenever myself/anyone else looks at this branch again: just realized it might be possible to add an attribute or some kind of meta to denote the style object as the "adopted stylesheet" internal node to fix the issues I was having. Then vs replace the shadowRoot always keep existing styles there and target the one to replace if possible (ref on whether data attribs are ok @ https://stackoverflow.com/a/48841488/1291659 )